requirements.txt
beautifulsoup4==4.8.1
flask==1.1.1
jinja2==2.10.3
jsonschema==3.1.1
openapi-spec-validator==0.2.8
pyyaml==5.1.2
requests==2.22.0
sqlalchemy==1.3.10
tox==3.14.0
